fix(#1877): move getVisibleLayerBounds out of BuilderMap.tsx
Round 2's fix imported getVisibleLayerBounds statically from
BuilderMap.tsx into use-builder-layers.ts, which is eagerly bundled.
That defeats MapBuilderPage.tsx's lazy(() => import(BuilderMap))
split: a production build reports INEFFECTIVE_DYNAMIC_IMPORT and folds
the map renderer, MapLibre GL and its worker/CSS side effects into the
eager entry chunk (BuilderMap's own chunk collapsed to a 0.07 kB stub;
MapBuilderPage grew from 334.56 kB to 351.60 kB gzip).

Move the pure bounds-merge helpers into their own module,
builder-bounds.ts, with no import of BuilderMap.tsx or anything else
eager, and import it from both BuilderMap.tsx and the hook. Confirmed
with npm run build: the warning is gone and BuilderMap is back to its
own 18.04 kB (6.73 kB gzip) chunk.
